Reduce warnings
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@21229 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ daq/rawdata/modules/src/DeSerializerPXD.cc
@@ -53,7 +53,7 @@ void DeSerializerPXDModule::initialize()
   }
 
   // Open receiver sockets
-  for (int i = 0; i < m_hosts.size(); i++) {
+  for (int i = 0; i < (int)(m_hosts.size()); i++) {
     m_recvs.push_back(new EvtSocketSend(m_hosts[i], m_ports[i]));
   }
